+++ MYPOLICY/app/src/main/java/com/example/mypolicy/adapter/SearchAdapter.java
@@ -44,7 +44,7 @@ public class SearchAdapter extends RecyclerView.Adapter<SearchViewHolder>{
     @Override
     public void onBindViewHolder(@NonNull SearchViewHolder holder, int position) {
 
-        final int pCode=sList.get(position).getP_code();
+        final long pCode=sList.get(position).getP_code();
         Log.d("피코드피코드피코드",""+pCode);
 
         holder.title.setText(sList.get(position).getTitle());//String이라 문제 없음
@@ -52,7 +52,7 @@ public class SearchAdapter extends RecyclerView.Adapter<SearchViewHolder>{
 
 
 
-        holder.p_code.setText(Integer.toString(pCode));//int를 String으로 파싱
+        holder.p_code.setText(Long.toString(pCode));//int를 String으로 파싱
         holder.category.setText(sList.get(position).getCategory());
 
         int matchScore=sList.get(position).getMatch_score();
