+++ MYPOLICY/app/src/main/java/com/example/mypolicy/model/StoreData.java
@@ -3,16 +3,16 @@ package com.example.mypolicy.model;
 import java.util.Date;
 
 public class StoreData {
-    private int p_code;
+    private long p_code;
     private String title;
     private Date apply_start;
     private Date apply_end;
 
-    public int getP_code() {
+    public long getP_code() {
         return p_code;
     }
 
-    public void setP_code(int p_code) {
+    public void setP_code(long p_code) {
         this.p_code = p_code;
     }
